+++ plugins/canvas/src/CanvasFeatureRenderer/index.ts
@@ -1,7 +1,4 @@
-
-
 export { default as ReactComponent } from './components/CanvasFeatureRendering'
 export { default as configSchema } from './configSchema'
 
-
-export {default} from './CanvasFeatureRenderer'
+export { default } from './CanvasFeatureRenderer'