+++ plugins/comparative-adapters/src/PairwiseIndexedPAFAdapter/configSchema.ts
@@ -16,7 +16,7 @@ const PairwiseIndexedPAFAdapter = ConfigurationSchema(
       type: 'stringArray',
       defaultValue: [],
       description:
-        'Array of assembly names to use for this file. The target assembly name is the first value in the array, query assembly name is the second',
+        'Array of assembly names to use for this file. The query assembly name is the first value in the array, target assembly name is the second',
     },
     /**
      * #slot