+++ products/jbrowse-cli/src/commands/create.ts
@@ -40,7 +40,7 @@ export default class Create extends JBrowseCommand {
       description:
         'Overwrites existing JBrowse 2 installation if present in path',
     }),
-    // will need to account for pagenation once there is a lot of releases
+    // will need to account for pagination once there is a lot of releases
     listVersions: Flags.boolean({
       char: 'l',
       description: 'Lists out all versions of JBrowse 2',